+++ frame/election-provider-multi-phase/test-staking-e2e/src/mock.rs
@@ -69,7 +69,7 @@ frame_support::construct_runtime!(
 
 pub(crate) type AccountId = u128;
 pub(crate) type AccountIndex = u32;
-pub(crate) type BlockNumber = u32;
+pub(crate) type BlockNumber = u64;
 pub(crate) type Balance = u64;
 pub(crate) type VoterIndex = u32;
 pub(crate) type TargetIndex = u16;
